changed syntax

+++ bin/haproxy_get_ssl.py
@@ -45,8 +45,9 @@ if check == "1":
         request = (
         "cd /opt/certs/letsencrypt && letsencrypt certonly --csr server.der --standalone --non-interactive --agree-tos --email %s --standalone-supported-challenges http-01" % email)
         os.system(request)
-        if !os.path.exists("/opt/certs/letsencrypt/0001_chain.pem"):
-            sys.exet(2)
+        cert_file = os.path.exists("/opt/certs/letsencrypt/0001_chain.pem")
+        if cert_file == False:
+            sys.exit(2)
         os.system(" cd /opt/certs/letsencrypt && cat 0001_chain.pem ../server.key > ../server.bundle.pem")
         os.system("rm -rf /opt/certs/letsencrypt")
         r_server.set("need_CSR", "0")
